Add config option to disable parts of management API
