+++ maubot/management/api/client_proxy.py
@@ -14,6 +14,7 @@
 # You should have received a copy of the GNU Affero General Public License
 # along with this program.  If not, see <https://www.gnu.org/licenses/>.
 from aiohttp import web, client as http
+
 from ...client import Client
 from .base import routes
 from .responses import resp
